[app_dart] move datastore/config to service/config (#1196)
