+++ app_dart/lib/src/request_handlers/check_for_waiting_pull_requests.dart
@@ -10,13 +10,12 @@ import 'package:github/github.dart';
 import 'package:graphql/client.dart';
 import 'package:meta/meta.dart';
 
-import '../datastore/config.dart';
 import '../foundation/providers.dart';
 import '../foundation/typedefs.dart';
 import '../request_handling/api_request_handler.dart';
 import '../request_handling/authentication.dart';
 import '../request_handling/body.dart';
-
+import '../service/config.dart';
 import 'check_for_waiting_pull_requests_queries.dart';
 import 'refresh_cirrus_status.dart';
