+++ app_dart/lib/src/request_handlers/get_build_status.dart
@@ -8,10 +8,10 @@ import 'package:cocoon_service/protos.dart' show BuildStatusResponse, EnumBuildS
 import 'package:cocoon_service/src/service/datastore.dart';
 import 'package:meta/meta.dart';
 
-import '../datastore/config.dart';
 import '../request_handling/body.dart';
 import '../request_handling/request_handler.dart';
 import '../service/build_status_provider.dart';
+import '../service/config.dart';
 
 @immutable
 class GetBuildStatus extends RequestHandler<Body> {
