+++ app_dart/lib/src/request_handling/api_request_handler.dart
@@ -9,7 +9,7 @@ import 'dart:typed_data';
 
 import 'package:meta/meta.dart';
 
-import '../datastore/config.dart';
+import '../service/config.dart';
 
 import 'authentication.dart';
 import 'body.dart';
