+++ app_dart/lib/src/service/access_token_provider.dart
@@ -5,8 +5,8 @@
 import 'package:googleapis_auth/auth_io.dart';
 import 'package:http/http.dart' as http;
 
-import '../datastore/config.dart';
 import '../model/appengine/service_account_info.dart';
+import 'config.dart';
 
 /// Function signature for a [TaskService] provider.
 typedef AccessTokenServiceProvider = AccessTokenService Function(Config config);
