+++ app_dart/test/src/request_handling/fake_authentication.dart
@@ -5,7 +5,7 @@
 import 'dart:io';
 
 import 'package:appengine/appengine.dart';
-import 'package:cocoon_service/src/datastore/config.dart';
+import 'package:cocoon_service/src/service/config.dart';
 import 'package:cocoon_service/src/foundation/typedefs.dart';
 import 'package:cocoon_service/src/model/appengine/key_helper.dart';
 import 'package:cocoon_service/src/request_handling/authentication.dart';
